+++ src/classic.ts
@@ -96,7 +96,7 @@ export const zostr = {
   nip05: {
     identifier: () => classicSchema(z.ZodString, nip05.nip05.identifier()),
     nostrJsonDocument: () =>
-      classicSchema(z.ZodObject, nip05.nip05.nostrJsonDocument()),
+      z.object(nip05.nip05.nostrJsonDocument()._zod.def.shape),
     formatIdentifier: nip05.nip05.formatIdentifier,
   },
 
@@ -121,6 +121,6 @@ export const zostr = {
   // NIP-11 relay information document
   nip11: {
     relayInformationDocument: () =>
-      classicSchema(z.ZodObject, nip11.nip11.relayInformationDocument()),
+      z.object(nip11.nip11.relayInformationDocument()._zod.def.shape),
   },
 };